// apps/end-user-app/src/types/finance.ts
export interface Asset {
  reference: string
  name: string
  symbol: string
  decimals: number
  is_settleable: boolean
}
// ...
export interface Instrument {
  reference: string
  name: string
  quantity_decimals: number
  px_decimals: number
  instrument_type: string
  base_asset_reference?: string
  quote_asset_reference?: string
  settlement_asset_reference?: string
  underlying_asset_reference?: string
  staking_asset_reference?: string
  yielding_asset_reference?: string
}

export interface CreateInstrumentFormInputs {
  name: string
  quantity_decimals: number
  px_decimals: number
  instrument_type: string
  base_asset_reference?: string
  quote_asset_reference?: string
  settlement_asset_reference?: string
  underlying_asset_reference?: string
  staking_asset_reference?: string
  yielding_asset_reference?: string
}

export interface UpdateInstrumentFormInputs {
  name?: string
  quantity_decimals?: number
  px_decimals?: number
  instrument_type?: string
  base_asset_reference?: string
  quote_asset_reference?: string
  settlement_asset_reference?: string
  underlying_asset_reference?: string
  staking_asset_reference?: string
  yielding_asset_reference?: string
}